[cmake][build] Move project sources to using Conan
