+++ CMakePresets.json
@@ -1,0 +1,42 @@
+{
+  "version": 6,
+  "include": ["./ConanPresets.json"],
+  "configurePresets": [
+    {
+      "name": "debug-conan",
+      "displayName": "Debug config (Conan)",
+      "description": "Debug config with tests (Conan dependencies used)",
+      "inherits": "conan-debug",
+      "installDir": "${sourceDir}/build/Debug/install",
+      "cacheVariables": {
+        "SCP_BUILD_TESTS": true
+      }
+    },
+    {
+      "name": "release-conan",
+      "displayName": "Release config (Conan)",
+      "description": "Release config (Conan dependencies used)",
+      "installDir": "${sourceDir}/build/Release/install",
+      "inherits": "conan-release"
+    },
+    {
+      "name": "release-with-tests-conan",
+      "displayName": "Release config with tests (Conan)",
+      "description": "Release config with tests (Conan dependencies used)",
+      "inherits": "release-conan",
+      "cacheVariables": {
+        "SCP_BUILD_TESTS": true
+      }
+    }
+  ],
+  "buildPresets": [
+    {
+      "name": "debug",
+      "configurePreset": "debug-conan"
+    },
+    {
+      "name": "release",
+      "configurePreset": "release-conan"
+    }
+  ]
+}
