+++ src/manager/agents/common-module/include/common/common_utils.hpp
