+++ src/manager/agents/common-module/include/common/repository_url_parser.hpp
@@ -7,7 +7,8 @@
 #pragma once
 
 #include <string>
-#include "constants/command_constants.hpp"
+
+#include "common/command_constants.hpp"
 
 class RepositoryUrlParser
 {
