+++ src/manager/agents/common-module/include/common/sc_component_manager_command.hpp
