+++ src/manager/agents/common-module/include/common/sc_component_manager_keynodes.hpp
