+++ src/manager/agents/init-module/CMakeLists.txt
@@ -1,8 +1,13 @@
-file(GLOB_RECURSE SOURCES "*.cpp" "*.hpp")
+file(GLOB SOURCES CONFIGURE_DEPENDS
+    "*.hpp" "*.cpp"
+    "*/*.hpp" "*/*.cpp"
+    "*/*/*.hpp" "*/*/*.cpp"
+)
 
 add_library(init-lib SHARED ${SOURCES})
 target_link_libraries(init-lib LINK_PUBLIC common-lib)
-target_include_directories(init-lib PUBLIC ${CMAKE_CURRENT_LIST_DIR} ${SC_CONFIG_PATH})
+target_include_directories(init-lib PRIVATE ${CMAKE_CURRENT_LIST_DIR})
+set_target_properties(init-lib PROPERTIES LIBRARY_OUTPUT_DIRECTORY ${SC_EXTENSIONS_DIRECTORY})
 
 if(${SC_CLANG_FORMAT_CODE})
     target_clangformat_setup(init-lib)
