+++ src/manager/agents/search-module/utils/sc_component_manager_command_search.cpp
@@ -4,12 +4,12 @@
  * (See accompanying file COPYING.MIT or copy at http://opensource.org/licenses/MIT)
  */
 
-#include <algorithm>
-
 #include "sc_component_manager_command_search.hpp"
 
-#include "module/keynodes/ScComponentManagerKeynodes.hpp"
-#include "module/utils/common_utils.hpp"
+#include <algorithm>
+
+#include <common/sc_component_manager_keynodes.hpp>
+#include <common/common_utils.hpp>
 
 using namespace common_utils;
