+++ src/manager/agents/search-module/utils/sc_component_manager_command_search.hpp
@@ -7,8 +7,9 @@
 #pragma once
 
 #include <sc-memory/sc_agent_context.hpp>
-#include "commands/sc_component_manager_command.hpp"
-#include "module/keynodes/ScComponentManagerKeynodes.hpp"
+
+#include "common/sc_component_manager_command.hpp"
+#include "common/sc_component_manager_keynodes.hpp"
 
 class ScComponentManagerCommandSearch : public ScComponentManagerCommand
 {
