+++ src/manager/console-interface/include/console-interface/sc_component_manager_command_parser.hpp
@@ -8,8 +8,8 @@
 
 #include <regex>
 
-#include "common-module/commands/sc_component_manager_command.hpp"
-#include "common-module/constants/command_constants.hpp"
+#include <common/sc_component_manager_command.hpp>
+#include <common/command_constants.hpp>
 
 class ScComponentManagerParser
 {
